Fixed the area approximation methods and added java doc
The right Riemann sum did not account for the area of the last rectangle
. This was fixed by changing "b" to "b+1" in the condition of the for
loop.

The trapezoidal sum followed the same issue as the right Riemann sum,
fixed in the same manner.
It also did not calculate the final trapezoid's area value correctly.
This was fixed by changing "x == b" to "x >= b" in the if statement's
condition.